Create codefile
In this project, you will develop a system that can monitor the fill level of a trash can in real-time and alert the maintenance personnel when the trash can is full and needs to be emptied. The system will use ultrasonic sensors to measure the fill level of the trash can and send the data to a microcontroller. The microcontroller will then send the data to a cloud-based server via Wi-Fi or GSM module. The server will process the data and send a notification to the maintenance personnel when the trash can is full. The maintenance personnel can then empty the trash can in a timely manner, reducing the risk of overflowing trash and improving sanitation.